encapsulate gcloud instance deletion as part of creation

+++ backend/scripts/delete-doughnut-app-instance.sh
@@ -1,1 +1,8 @@
+#/bin/sh
+RUNING_APP_INSTANCE_COUNT=$(gcloud compute instances list --filter='tags:app-server' | grep RUNNING | wc -l | xargs)
+
+if [ ${RUNING_APP_INSTANCE_COUNT} -eq 0 ]; then
+	exit 0
+fi
+
 gcloud compute instances delete doughnut-app-instance --quiet --zone us-east1-b
